+++ src/pages/Coin.tsx
@@ -7,7 +7,7 @@ import {
   useParams,
 } from "react-router-dom";
 import styled from "styled-components";
-import { fetchCoinInfo, fetchCoinTickers } from "../api";
+import { fetchCoinInfo, fetchCoinTickers } from "../api/api";
 import { Helmet } from "react-helmet";
 import Controller from "../components/Controller";
