+++ src/pages/Coins.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { Link } from "react-router-dom";
 import { useQuery } from "react-query";
-import { fetchCoins } from "../api";
+import { fetchCoins } from "../api/api";
 import { Helmet } from "react-helmet";
 import Controller from "../components/Controller";
